% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alaska_sponge_coral_fish}
\alias{alaska_sponge_coral_fish}
\title{Data to analyze sponge-coral-fish associations}
\usage{
data(alaska_sponge_coral_fish)
}
\description{
Data used in
\href{https://onlinelibrary.wiley.com/doi/10.1111/geb.70035}{Thorson et al. 2025}
and copied from \href{https://zenodo.org/records/15001896}{Zenodo} to allow
inclusion as a package vignette.  This includes bottom-trawl samples of rockfish
and flatfish from the Alaska Fisheries Science Center in the Gulf of Alaska
and Aleutian Islands, and drop-camera samples of rockfish, flatfish, sponges,
and corals from the Alaska Fisheries Science Center and
funded by the Deep Sea Coral Initiative.
}
\keyword{data}
