% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlantic_yellowtail}
\alias{atlantic_yellowtail}
\title{Northwest Atlantic yellowtail}
\usage{
data(atlantic_yellowtail)
}
\description{
Data used to demonstrate and test seasonal spatio-temporal
index standardization.
Data are from \doi{10.1093/icesjms/fsaa074}
}
\keyword{data}
