% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bering_sea}
\alias{bering_sea}
\title{Survey domain for the eastern and northern Bering Sea surveys}
\usage{
data(bering_sea)
}
\description{
Shapefile defining the spatial domain for the eastern and northern
Bering Sea bottom trawl surveys.
}
\keyword{data}
