% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bering_sea_capelin_forecasts}
\alias{bering_sea_capelin_forecasts}
\title{Data to demonstrate probabilistic forecasting}
\usage{
data(bering_sea_capelin_forecasts)
}
\description{
Data to estimate probabilistic forecasting, which bridges uncertainty
between short-term (decadal) projections and long-term (end-of-century)
forecasts.  Here we use capelin sampling from a survey trawl survey
in the eastern and northern Bering Sea, as well as annual
sea surface temperature anomalies.  The data set includes data frames
for both the fitted data \code{Data} and the projected values \code{New_Data}
}
\keyword{data}
