% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bering_sea_pollock_vast}
\alias{bering_sea_pollock_vast}
\title{Estimated proportion-at-age for Alaska pollock using VAST}
\usage{
data(bering_sea_pollock_vast)
}
\description{
Estimated proporrtion-at-age for Alaska pollock using the
package VAST, for comparison with output using tinyVAST.
}
\keyword{data}
