% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{classify_variables}
\alias{classify_variables}
\title{Classify variables path}
\usage{
classify_variables(model)
}
\arguments{
\item{model}{syntax for structural equation model}
}
\value{
Tagged-list defining exogenous and endogenous variables
}
\description{
\code{classify_variables} is copied from \code{sem:::classifyVariables}
}
\details{
Copied from package \code{sem} under licence GPL (>= 2) with permission from John Fox
}
