% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{condition_and_density}
\alias{condition_and_density}
\title{Condition and density example}
\usage{
data(condition_and_density)
}
\description{
Data used to demonstrate and test a bivariate model for morphometric condition
(i.e., residuals in a weight-at-length relationship) and density for fishes, using
the same example as was provided as a wiki example for VAST.
Data are from \doi{10.3354/meps13213}
}
\keyword{data}
