% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Families}
\alias{Families}
\alias{lognormal}
\alias{tweedie}
\title{Additional families}
\usage{
lognormal(link = "log")

tweedie(link = "log")
}
\arguments{
\item{link}{Link.}
}
\value{
A list with elements common to standard R family objects including \code{family},
\code{link}, \code{linkfun}, and \code{linkinv}.
}
\description{
Additional families compatible with \code{\link[=tinyVAST]{tinyVAST()}}.
}
\details{
Copied from sdmTMB
}
\examples{
lognormal(link = "log")
tweedie(link = "log")
}
