% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_data.tinyVAST}
\alias{get_data.tinyVAST}
\title{Get data}
\usage{
\method{get_data}{tinyVAST}(x, ...)
}
\arguments{
\item{x}{output from \code{\link[=tinyVAST]{tinyVAST()}}}

\item{...}{not used}
}
\description{
S3 generic from package insight, used for crossvalidation
}
