% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetResponse.tinyVAST}
\alias{GetResponse.tinyVAST}
\title{Get response}
\usage{
\method{GetResponse}{tinyVAST}(model, ...)
}
\arguments{
\item{model}{output from \code{\link[=tinyVAST]{tinyVAST()}}}

\item{...}{not used}
}
\description{
S3 generic from package cv, used for crossvalidation
}
