% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_domains.R
\name{make_nngp_domain}
\alias{make_nngp_domain}
\title{Make spatial object for NNGP}
\usage{
make_nngp_domain(sf_areal, nn)
}
\arguments{
\item{sf_areal}{an sf or sfc object containing "POLYGON" and "MULTIPOLYGON" types}

\item{nn}{Number of nearest neighbors}
}
\value{
a list of NNGP data objects created by tinyVAST:::make_nngp_data,
as well as the provided sf object
}
\description{
Convert an sf areal model object to a class \code{"nngp_domain"} that is recognized
by tinyVAST for nearest neighbors Gaussian process models
}
