% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sem_ram.R
\name{make_sem_ram}
\alias{make_sem_ram}
\title{Make a RAM (Reticular Action Model) from a SEM (structural equation model)}
\usage{
make_sem_ram(sem, variables, quiet = FALSE, covs = variables)
}
\arguments{
\item{sem}{structural equation model structure, passed to either \code{specifyModel}
or \code{specifyEquations} and then parsed to control
the set of path coefficients and variance-covariance parameters}

\item{variables}{A character vector listing the set of variables}

\item{quiet}{Boolean indicating whether to print messages to terminal}

\item{covs}{A character vector listing variables for which to estimate a standard deviation}
}
\value{
An S3-class \code{"sem_ram"} containing:
\describe{
\item{\code{model}}{Output from \code{\link[sem]{specifyEquations}} or \code{\link[sem]{specifyModel}}
that defines paths and parameters}
\item{\code{ram}}{reticular action module (RAM) describing dependencies}
}
}
\description{
\code{make_sem_ram} converts SEM arrow notation to \code{ram} describing SEM parameters
}
