% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{parse_path}
\alias{parse_path}
\title{Parse path}
\usage{
parse_path(path)
}
\arguments{
\item{path}{text to parse}
}
\value{
Tagged-list defining variables and direction for a specified path coefficient
}
\description{
\code{parse_path} is copied from \code{sem::parse.path}
}
\details{
Copied from package \code{sem} under licence GPL (>= 2) with permission from John Fox
}
