% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tinyVAST}
\alias{print.tinyVAST}
\title{print summary of tinyVAST model}
\usage{
\method{print}{tinyVAST}(x, ...)
}
\arguments{
\item{x}{output from \code{tinyVAST}}

\item{...}{not used}
}
\value{
invisibly returns a named list of key model outputs and summary
statements
}
\description{
print summary of tinyVAST model
}
