% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{red_snapper}
\alias{red_snapper}
\title{Presence/absence, count, and biomass data for red snapper}
\usage{
data(red_snapper)
}
\description{
Data used to demonstrate and test analysis using multiple data types
}
\keyword{data}
