% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{red_snapper_shapefile}
\alias{red_snapper_shapefile}
\title{Shapefile for red snapper analysis}
\usage{
data(red_snapper_shapefile)
}
\description{
Spatial extent used for red snapper analysis, derived from Chap-7 of \doi{10.1201/9781003410294}
}
\keyword{data}
