% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{lognormal}
\alias{nbinom2}
\alias{nbinom1}
\alias{tweedie}
\alias{student}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sdmTMB}{\code{\link[sdmTMB:families]{lognormal}}, \code{\link[sdmTMB:families]{nbinom1}}, \code{\link[sdmTMB:families]{nbinom2}}, \code{\link[sdmTMB:families]{student}}, \code{\link[sdmTMB:families]{tweedie}}}
}}

