% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reload_model}
\alias{reload_model}
\title{Reload a previously fitted model}
\usage{
reload_model(x, check_gradient = TRUE)
}
\arguments{
\item{x}{Output from \code{\link{tinyVAST}}, potentially with DLLs not linked}

\item{check_gradient}{Whether to check the gradients of the reloaded model}
}
\value{
Output from \code{\link{tinyVAST}} with DLLs relinked
}
\description{
\code{reload_model} allows a user to save a fitted model, reload it in a new
R terminal, and then relink the DLLs so that it functions as expected.
}
