% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{residuals.tinyVAST}
\alias{residuals.tinyVAST}
\title{Calculate deviance or response residuals for tinyVAST}
\usage{
\method{residuals}{dsem}(object, type = c("deviance", "response"), ...)
}
\arguments{
\item{object}{Output from \code{\link{fit}}}

\item{type}{which type of residuals to compute (only option is \code{"deviance"} or \code{"response"} for now)}

\item{...}{Note used}
}
\description{
Calculate residuals
}
