% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salmon_returns}
\alias{salmon_returns}
\title{North Pacific salmon returns}
\usage{
data(salmon_returns)
}
\description{
Data used to demonstrate and test multivariate second-order autoregressive
models using a simultaneous autoregressive (SAR) process across regions.
Data are from \doi{10.1002/mcf2.10023}
}
\keyword{data}
