% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity.R
\name{sanity}
\alias{sanity}
\title{Sanity check of a tinyVAST model}
\usage{
sanity(object, big_sd_log10 = 2, gradient_thresh = 0.001, silent = FALSE)
}
\arguments{
\item{object}{Fitted model from \code{\link[=tinyVAST]{tinyVAST()}}.}

\item{big_sd_log10}{Value to check size of standard errors against. A value
of 2 would indicate that standard errors greater than \code{10^2} (i.e., 100)
should be flagged.}

\item{gradient_thresh}{Gradient threshold to issue warning.}

\item{silent}{Logical: suppress messages? Useful to set to \code{TRUE} if running
large numbers of models and just interested in returning sanity list
objects.}
}
\value{
An invisible named list of checks.
}
\description{
Sanity check of a tinyVAST model
}
