% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sea_ice}
\alias{sea_ice}
\title{Arctic September sea ice concentrations}
\usage{
data(sea_ice)
}
\description{
Data used to demonstrate and test empirical orthogonal function
generalized linear latent variable model (EOF-GLLVM)
}
\keyword{data}
