% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{simulate_conditional_gmrf}
\alias{simulate_conditional_gmrf}
\title{Conditional simulation from a GMRF}
\usage{
simulate_conditional_gmrf(Q, observed_idx, x_obs, n_sims = 1)
}
\arguments{
\item{Q}{precision for a zero-centered GMRF.}

\item{observed_idx}{integer vector listing rows of \code{Q} corresponding to
fixed measurements}

\item{x_obs}{numeric vector with fixed values for indices \code{observed_idx}}

\item{n_sims}{integer listing number of simulated values}
}
\value{
A matrix with \code{n_sims} columns and a row for every row of \code{Q} not in
\code{observed_idx}, with simulations for those rows
}
\description{
Generates samples from a Gaussian Markov random field (GMRF) conditional upon
fixed values for some elements.
}
