% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{spatial_cor}
\alias{spatial_cor}
\title{Approximate spatial correlation}
\usage{
spatial_cor(Q, mesh, coord, pred)
}
\arguments{
\item{Q}{sparse precision matrix}

\item{mesh}{SPDE mesh}

\item{coord}{vector of length-2 with spatial coordinates for focal point}

\item{pred}{matrix with two columns and multiple rows, with location
for points to predict correlation}
}
\value{
A vector with length \code{nrow(pred)} giving the spatial correlation
}
\description{
Extract the approximated spatial correlation between one coordinate
and other coordinates using a sparse precision and SPDE mesh
}
