\name{tinyVAST-package}
\alias{tinyVAST-package}
\alias{tinyVAST}
\docType{package}
\title{
\packageTitle{tinyVAST}
}
\description{
\packageDescription{tinyVAST}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tinyVAST}
\packageIndices{tinyVAST}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tinyVAST}

Maintainer: \packageMaintainer{tinyVAST}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
