% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{vcov.tinyVAST}
\alias{vcov.tinyVAST}
\title{Extract Variance-Covariance Matrix}
\usage{
\method{vcov}{tinyVAST}(object, which = c("fixed", "random", "both"), ...)
}
\arguments{
\item{object}{output from \code{\link[=tinyVAST]{tinyVAST()}}}

\item{which}{whether to extract the covariance among fixed effects, random effects, or both}

\item{...}{ignored, for method compatibility}
}
\value{
A square matrix containing the estimated covariances among the parameter estimates in the model.
The dimensions dependend upon the argument \code{which}, to determine whether fixed, random effects,
or both are outputted.
}
\description{
extract the covariance of fixed effects, or both fixed and random effects.
}
