% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{add_predictions}
\alias{add_predictions}
\title{Add predictions to data-list}
\usage{
add_predictions(object, newdata)
}
\arguments{
\item{object}{Output from [fit()].}

\item{newdata}{New data-frame of independent variables used to predict the response.}
}
\description{
Given user-provided `newdata`, expand the object `tmb_data`
   to include predictions corresponding to those new observations
}
