% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{deviance_explained}
\alias{deviance_explained}
\title{Calculate deviance explained}
\usage{
deviance_explained(x, null_formula)
}
\arguments{
\item{x}{output from \verb{\code{tinyVAST()}}}

\item{null_formula}{formula for the null model.  If missing, it uses
\code{null_formula = response ~ 1}. For multivariate models, it
might make sense to use \code{null_formula = response ~ category}}
}
\description{
\code{deviance_explained} fits a null model, calculates the deviance relative to
a saturated model for both the original and the null model, and uses these
to calculate the proportion of deviance explained.
}
