% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{Families}
\alias{Families}
\alias{lognormal}
\alias{tweedie}
\alias{delta_lognormal}
\title{Additional families}
\usage{
lognormal(link = "log")

tweedie(link = "log")

delta_lognormal(link1 = "logit", link2 = "log")
}
\arguments{
\item{link}{Link.}

\item{link1}{First linear predictor link}

\item{link2}{Second linear predictor link}
}
\value{
A list with elements common to standard R family objects including \code{family},
\code{link}, \code{linkfun}, and \code{linkinv}.
}
\description{
Additional families compatible with \pkg{tinyVAST}.
}
\details{
Copied from sdmTMB
}
\examples{
lognormal(link = "log")
tweedie(link = "log")
delta_lognormal()
}
