% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit vector autoregressive spatio-temporal model}
\usage{
fit(
  data,
  formula,
  sem = NULL,
  estimate_delta0 = FALSE,
  family_link = c(0, 0),
  data_colnames = list(spatial = c("x", "y"), variable = "var", time = "time",
    distribution = "dist"),
  times = sort(unique(data[, data_colnames$time])),
  variables = unique(data[, data_colnames$variable]),
  spatial_graph = NULL,
  control = tinyVASTcontrol(),
  ...
)
}
\arguments{
\item{data}{Data-frame of predictor and response variables.}

\item{formula}{Formula with response on left-hand-side and predictors on right-hand-side,
parsed by \code{mgcv} and hence allowing \code{s(.)} for splines}

\item{sem}{Specification for time-series structural equation model structure
including lagged or simultaneous effects.  See Details section in
\code{\link[dsem]{make_ram}} for more description}

\item{estimate_delta0}{Boolean indicating whether to estimate deviations from equilibrium in initial year
as fixed effects, or alternatively to assume that dynamics start at some stochastic draw away from
the stationary distribution}

\item{family_link}{Vector of length-two, indicating the distribution and link-function}

\item{times}{A character vector listing the set of times in order}

\item{variables}{A character vector listing the set of variables}

\item{spatial_graph}{Object that represents spatial relationships, either using \code{fmesher}
to apply the SPDE method, \code{igraph} to apply a simultaneous autoregressive (SAR)
process, or \code{NULL} to specify a single site.}

\item{control}{Output from \code{\link{tinyVASTcontrol}}, used to define user
settings, and see documentation for that function for details.}

\item{...}{Additional parameters passed to \code{\link[dsem]{fit_tmb}}}
}
\description{
Fits a vector autoregressive spatio-temporal model using
a minimal feature-set, and widely used interface for objects
}
\details{
\link{tinyVAST} includes four basic inputs that specify the model structure:
\itemize{
\item \link{formula} specifies covariates and splines in a Generalized Additive Model;
\item \link{sem} specifies interactions among variables and over time
\item \link{spatial_graph} specifies spatial correlations
}

the default \link{sem=NULL} turns off all multivariate and temporal indexing, such
that \code{spatial_graph} is then ignored, and the model collapses
to a standard model using \code{mgcv::gam}.  To specify a univeriate spatial model,
the user must specify both \code{spatial_graph} and \code{sem=""}, where the latter
is then parsed to include a single exogenous variance for the single variable\tabular{ll}{
   Model type \tab How to specify \cr
   Spatial model \tab specify both \code{spatial_graph} and \code{sem=""}, where the latter is then parsed to include a single exogenous variance for the single variable \cr
}
}
\examples{
methods(class="tinyVAST")

}
