% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit vector autoregressive spatio-temporal model}
\usage{
fit(
  data,
  formula,
  sem = NULL,
  estimate_delta0 = FALSE,
  data_colnames = list(spatial = c("x", "y"), variable = "var", time = "time"),
  times = sort(unique(data[, data_colnames$time])),
  variables = unique(data[, data_colnames$variable]),
  spatial_graph = NULL,
  quiet = FALSE
)
}
\description{
Fits a vector autoregressive spatio-temporal model using
 a minimal feature-set, and widely used interface for objects
}
\examples{
methods(class="tinyVAST")

}
