% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit vector autoregressive spatio-temporal model}
\usage{
fit(data, formula)
}
\description{
Fits a vector autoregressive spatio-temporal model using
 a minimal feature-set, and widely used interface for objects
}
