% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit vector autoregressive spatio-temporal model}
\usage{
fit(
  data,
  formula,
  sem = NULL,
  dsem = NULL,
  family = list(obs = gaussian()),
  data_colnames = list(spatial = c("x", "y"), variable = "var", time = "time",
    distribution = "dist"),
  times = seq(min(data[, data_colnames$time]), max(data[, data_colnames$time])),
  variables = unique(data[, data_colnames$variable]),
  spatial_graph = NULL,
  control = tinyVASTcontrol(),
  ...
)
}
\arguments{
\item{data}{Data-frame of predictor and response variables.}

\item{formula}{Formula with response on left-hand-side and predictors on right-hand-side,
parsed by \code{mgcv} and hence allowing \code{s(.)} for splines or \code{offset(.)} for
an offset.}

\item{sem}{Specification for structural equation model structure for
constructing a space-variable interaction.
See code{\link{make_sem_ram}} for more description.}

\item{dsem}{Specification for time-series structural equation model structure
including lagged or simultaneous effects for
constructing a space-variable interaction.  See
\code{\link{make_dsem_ram}}  or \code{\link{make_eof_ram}}
for more description}

\item{family}{named list of family and link functions, with names corresponding to levels
of \code{data$dist}}

\item{times}{A character vector listing the set of times in order}

\item{variables}{A character vector listing the set of variables}

\item{spatial_graph}{Object that represents spatial relationships, either using \code{fmesher}
to apply the SPDE method, \code{igraph} to apply a simultaneous autoregressive (SAR)
process, or \code{NULL} to specify a single site.}

\item{control}{Output from \code{\link[=tinyVASTcontrol]{tinyVASTcontrol()}}, used to define user
settings, and see documentation for that function for details.}
}
\description{
Fits a vector autoregressive spatio-temporal model using
a minimal feature-set, and widely used interface for objects
}
\details{
\code{tinyVAST} includes four basic inputs that specify the model structure:
\itemize{
\item \code{formula} specifies covariates and splines in a Generalized Additive Model;
\item \code{dsem} specifies interactions among variables and over time, constructing
the space-time-variable interaction.
\item \code{sem} specifies interactions among variables and over time, constructing
the space-variable interaction.
\item \code{spatial_graph} specifies spatial correlations
}

the default \code{dsem=NULL} turns off all multivariate and temporal indexing, such
that \code{spatial_graph} is then ignored, and the model collapses
to a standard model using \code{mgcv::gam}.  To specify a univeriate spatial model,
the user must specify both \code{spatial_graph} and \code{dsem=""}, where the latter
is then parsed to include a single exogenous variance for the single variable\tabular{ll}{
   \strong{Model type} \tab \strong{How to specify} \cr
   Generalized additive model \tab specify \code{spatial_graph=NULL} and \code{dsem=""}, and then use \code{formula} to specify splines and covariates \cr
   Dynamic structural equation model (including vector autoregressive, dynamic factor analysis, ARIMA, and structural equation models) \tab specify \code{spatial_graph=NULL} and use \code{dsem} to specify interactions among variables and over time \cr
   Univeriate spatial model \tab specify \code{spatial_graph} and \code{dsem=""}, where the latter is then parsed to include a single exogenous variance for the single variable \cr
   Multivariate spatial model \tab specify \code{spatial_graph} and use \code{dsem} (without any lagged effects) to specify spatial interactions \cr
   Vector autoregressive spatio-temporal model \tab specify \code{spatial_graph} and use \code{dsem=""} to specify interactions among variables and over time, where spatio-temporal variables are constructed via the separable interaction of \code{dsem} and \code{spatial_graph} \cr
}
}
\examples{
# Simulate a 2D AR1 spatial process with a cyclic confounder w
n_x = n_y = 25
n_w = 10
R_xx = exp(-0.4 * abs(outer(1:n_x, 1:n_x, FUN="-")) )
R_yy = exp(-0.4 * abs(outer(1:n_y, 1:n_y, FUN="-")) )
z = mvtnorm::rmvnorm(1, sigma=kronecker(R_xx,R_yy) )

# Simulate nuissance parameter z from oscillatory (day-night) process
w = sample(1:n_w, replace=TRUE, size=length(z))
Data = data.frame( expand.grid(x=1:n_x, y=1:n_y), w=w, z=as.vector(z) + cos(w/n_w*2*pi))
Data$n = Data$z + rnorm(nrow(Data), sd=1)

# Add columns for multivariate and temporal dimensions
Data$var = "n"
Data$time = 2020

# make mesh
mesh = fmesher::fm_mesh_2d( Data[,c('x','y')], n=100 )

# fit model
out = fit( data = Data,
           formula = n ~ s(w),
           spatial_graph = mesh,
           control = tinyVASTcontrol(quiet=TRUE, trace=0),
           sem = "" )

}
