% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{integrate_output}
\alias{integrate_output}
\title{Monte Carlo integration for abundance}
\usage{
integrate_output(
  object,
  newdata,
  bias.correct = TRUE,
  apply.epsilon = FALSE,
  intern = FALSE,
  W_gz,
  V_gz,
  area
)
}
\arguments{
\item{object}{Output from \code{\link[=fit]{fit()}}.}

\item{newdata}{New data-frame of independent variables used to predict the response.}

\item{bias.correct}{logical indicating if bias correction should be applied}

\item{area}{value used for area-weighted expansion of estimated density surface
for each row of \code{newdata}.}
}
\description{
Applies Monte Carlo integration to approximate area-expanded abundance
}
