% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{integrate_output}
\alias{integrate_output}
\title{Monte Carlo integration for abundance}
\usage{
integrate_output(
  object,
  newdata,
  V_gz,
  area,
  W_gz,
  bias.correct = TRUE,
  apply.epsilon = FALSE,
  intern = FALSE
)
}
\arguments{
\item{object}{Output from \code{\link[=fit]{fit()}}.}

\item{newdata}{New data-frame of independent variables used to predict the response.}

\item{V_gz}{Settings for expansion.}

\item{area}{value used for area-weighted expansion of estimated density surface
for each row of \code{newdata}.}

\item{W_gz}{Covariates for expansion.}

\item{bias.correct}{logical indicating if bias correction should be applied}

\item{apply.epsilon}{Apply epsilon bias correction?}

\item{intern}{Do Laplace approximation on C++ side? Passed to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.}
}
\description{
Applies Monte Carlo integration to approximate area-expanded abundance
}
