% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{logLik.tinyVAST}
\alias{logLik.tinyVAST}
\title{Extract the (marginal) log-likelihood of a tinyVAST model}
\usage{
\method{logLik}{tinyVAST}(object, ...)
}
\value{
object of class \code{logLik} with attributes
\item{val}{log-likelihood}
\item{df}{number of parameters}
}
\description{
Extract the (marginal) log-likelihood of a tinyVAST model
}
