% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_domains.R
\name{plot.nngp_domain}
\alias{plot.nngp_domain}
\title{Plot nngp_domain}
\usage{
\method{plot}{nngp_domain}(x, y, ...)
}
\arguments{
\item{x}{output from \code{make_nngp_domain}}

\item{y}{not used}

\item{...}{not used}
}
\description{
Plot nngp_domain
}
