% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tinyVAST}
\alias{predict.tinyVAST}
\title{Predict using vector autoregressive spatio-temporal model}
\usage{
\method{predict}{tinyVAST}(object, newdata, ...)
}
\description{
Predicts values given new covariates using a tinyVAST model
}
