% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tinyVAST}
\alias{predict.tinyVAST}
\title{Predict using vector autoregressive spatio-temporal model}
\usage{
\method{predict}{tinyVAST}(object, newdata, ...)
}
\arguments{
\item{object}{Output from \code{\link[=fit]{fit()}}.}

\item{newdata}{New data-frame of independent variables used to predict the response.}

\item{...}{Not used.}
}
\description{
Predicts values given new covariates using a tinyVAST model
}
