% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tinyVAST}
\alias{predict.tinyVAST}
\title{Predict using vector autoregressive spatio-temporal model}
\usage{
\method{predict}{tinyVAST}(
  object,
  newdata,
  remove_origdata = FALSE,
  what = c("mu_g", "p_g", "palpha_g", "pgamma_g", "pepsilon_g", "pomega_g"),
  se.fit = FALSE,
  ...
)
}
\arguments{
\item{object}{Output from \code{\link[=tinyVAST]{tinyVAST()}}.}

\item{newdata}{New data-frame of independent variables used to predict the response.}

\item{remove_origdata}{Whether to eliminate the original data
from the TMB object, thereby speeding up the TMB object construction.  However, this
also eliminates information about random-effect variance, and is not
appropriate when requesting predictive standard errors or epsilon
bias-correction.}

\item{what}{What REPORTed object to output, where \code{p_g} is the linear
predictor, \code{mu_g} is the inverse-linked transformed predictor.
and others are additive components of the linear predictor.}

\item{se.fit}{Calculate standard errors?}

\item{...}{Not used.}
}
\description{
Predicts values given new covariates using a \pkg{tinyVAST} model
}
