% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project tinyVAST to future times (EXPERIMENTAL)}
\usage{
project(object, extra_times, newdata, what = "mu_g", future_var = TRUE)
}
\arguments{
\item{object}{fitted model from \code{tinyVAST(.)}}

\item{extra_times}{a vector of extra times, matching values in \code{newdata}}

\item{newdata}{data frame including new values for \code{time_variable}}

\item{what}{What REPORTed object to output, where
\code{mu_g} is the inverse-linked transformed predictor including both linear components,
\code{p_g} is the first linear predictor,
\code{palpha_g} is the first predictor from fixed covariates in \code{formula},
\code{pgamma_g} is the first predictor from random covariates in \code{formula} (e.g., splines),
\code{pomega_g} is the first predictor from spatial variation,
\code{pepsilon_g} is the first predictor from spatio-temporal variation,
\code{pxi_g} is the first predictor from spatially varying coefficients,
\code{p2_g} is the second linear predictor,
\code{palpha2_g} is the second predictor from fixed covariates in \code{formula},
\code{pgamma2_g} is the second predictor from random covariates in \code{formula} (e.g., splines),
\code{pomega2_g} is the second predictor from spatial variation,
\code{pepsilon2_g} is the second predictor from spatio-temporal variation, and
\code{pxi2_g} is the second predictor from spatially varying coefficients.}

\item{future_var}{logical indicating whether to simulate future process errors
from GMRFs, or just compute the predictive mean}
}
\value{
A vector of values corresponding to rows in \code{newdata}
}
\description{
Projects a fitted model forward in time.
}
