% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{red_grouper_diet}
\alias{red_grouper_diet}
\title{Data to demonstrate model-based diet proportions}
\usage{
data(red_grouper_diet)
}
\description{
Data to estimate predator-expanded stomach contents (PESC), i.e.,
a multi-prey single-predator model for stomach contents of red grouper
in the Gulf of Mexico that also estimates biomass density for red grouper.
Diet proportions are then the product of predicted diet proportions and prey
specific consumption, normalized across prey categories. Copied from
VAST data set \code{PESC_example_red_grouper}
}
\keyword{data}
