% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{lognormal}
\alias{nbinom2}
\alias{nbinom1}
\alias{tweedie}
\alias{student}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sdmTMB}{\code{\link[sdmTMB:lognormal]{lognormal()}}, \code{\link[sdmTMB:nbinom1]{nbinom1()}}, \code{\link[sdmTMB:nbinom2]{nbinom2()}}, \code{\link[sdmTMB:student]{student()}}, \code{\link[sdmTMB:tweedie]{tweedie()}}}
}}

