% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.tinyVAST}
\alias{residuals.tinyVAST}
\title{Calculate deviance or response residuals for tinyVAST}
\usage{
\method{residuals}{tinyVAST}(object, type = c("deviance", "response"), ...)
}
\arguments{
\item{object}{Output from \code{\link[=tinyVAST]{tinyVAST()}}}

\item{type}{which type of residuals to compute (only option is \code{"deviance"} or \code{"response"} for now)}

\item{...}{Note used}
}
\value{
a vector residuals, associated with each row of \code{data} supplied during fitting
}
\description{
Calculate residuals
}
