% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rmvnorm_prec}
\alias{rmvnorm_prec}
\title{Multivariate Normal Random Deviates using Sparse Precision}
\usage{
rmvnorm_prec(prec, n = 1, mu = rep(0, nrow(prec)))
}
\arguments{
\item{prec}{sparse precision (inverse-covariance) matrix.}

\item{n}{number of observations.}

\item{mu}{mean vector.}
}
\value{
a matrix with dimension \code{length(mu)} by
\code{n}, containing realized draws from the specified
mean and precision
}
\description{
This function provides a random number generator for
the multivariate normal distribution with mean equal
to \code{mu} and sparse precision matrix \code{prec}.
}
