% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rmvnorm_prec}
\alias{rmvnorm_prec}
\title{Multivariate Normal Random Deviates using Sparse Precision}
\usage{
rmvnorm_prec(n, Q, mean = rep(0, nrow(Q)))
}
\arguments{
\item{n}{number of observations.}

\item{Q}{sparse precision (inverse-covariance) matrix.}

\item{mean}{mean vector.}
}
\value{
a matrix with dimension \code{length(mean)} by
\code{n}, containing realized draws from the specified
mean and precision
}
\description{
This function provides a random number generator for
the multivariate normal distribution with mean equal
to \code{mean} and sparse precision matrix \code{Q}.
}
