% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetwork.R
\name{sfnetwork_evaluator}
\alias{sfnetwork_evaluator}
\title{Construct projection matrix for stream network}
\usage{
sfnetwork_evaluator(stream, loc, tolerance = 0.01)
}
\arguments{
\item{stream}{sfnetworks object representing stream network}

\item{loc}{sf object representing points to which are being projected}

\item{tolerance}{error-check tolerance}
}
\description{
Make sparse matrix to project from stream-network nodes
to user-supplied points
}
