% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetwork.R
\name{sfnetwork_mesh}
\alias{sfnetwork_mesh}
\title{Make mesh for stream network}
\usage{
sfnetwork_mesh(stream)
}
\arguments{
\item{stream}{sfnetworks object representing stream network}
}
\description{
Make mesh for stream network
}
