% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetwork.R
\name{simulate_sfnetwork}
\alias{simulate_sfnetwork}
\title{Simulate GMRF for stream network}
\usage{
simulate_sfnetwork(sfnetwork_mesh, theta, n = 1, what = c("samples", "Q"))
}
\arguments{
\item{sfnetwork_mesh}{Output from \code{\link{sfnetwork_mesh}}}

\item{theta}{Decorrelation rate}

\item{n}{number of simulated GMRFs}

\item{what}{Whether to return the simulated GMRF or its precision matrix}
}
\description{
Simulate GMRF for stream network
}
