% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{tinyVASTcontrol}
\alias{tinyVASTcontrol}
\title{Control parameters for tinyVAST}
\usage{
tinyVASTcontrol(
  opt_loops = 1,
  newton_loops = 0,
  eval.max = 10000,
  iter.max = 10000,
  getsd = TRUE,
  silent = getOption("tinyVAST.silent", TRUE),
  trace = getOption("tinyVAST.trace", 0),
  verbose = getOption("tinyVAST.verbose", FALSE),
  profile = c(),
  tmb_par = NULL,
  tmb_map = NULL,
  tmb_random = NULL,
  gmrf_parameterization = c("separable", "projection"),
  reml = FALSE,
  getJointPrecision = FALSE,
  calculate_deviance_explained = TRUE,
  run_model = TRUE,
  optimizer = c("nlminb", "L-BFGS-B"),
  suppress_user_warnings = FALSE,
  get_rsr = FALSE,
  extra_reporting = FALSE,
  use_anisotropy = FALSE,
  sar_adjacency = "queen",
  nearest_neighbors = NULL,
  barrier_stiffness = 0.01
)
}
\arguments{
\item{opt_loops}{Integer number of times to call nonlinear optimizer.}

\item{newton_loops}{Integer number of Newton steps to do after running
\code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{eval.max}{Maximum number of evaluations of the objective function
allowed. Passed to \code{control} in \code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{iter.max}{Maximum number of iterations allowed. Passed to \code{control} in
\code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{getsd}{Boolean indicating whether to call \code{\link[TMB:sdreport]{TMB::sdreport()}}}

\item{silent}{Disable terminal output for inner optimizer?}

\item{trace}{Parameter values are printed every \code{trace} iteration
for the outer optimizer. Passed to
\code{control} in \code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{verbose}{Output additional messages about model steps during fitting?}

\item{profile}{Character-vector passed to \link[TMB:MakeADFun]{TMB::MakeADFun} and see description
there.  Fixed effects that are highly correlated with random effects
can often be estimated faster (i.e., with fewer iterations) by adding
them to \code{profile}. The most common use-case is
\code{profile = c("alpha_j","alpha2_j")}.  However, doing so will have
a small impact on model estimates and predictions.}

\item{tmb_par}{named list of parameters used as starting values.  Elements that
have names that match those constructed internally then replace the
internally constructed starting values.  Those that match must have identical
shape to \code{tinyVAST(...)$internal$parlist}}

\item{tmb_map}{input passed to \link[TMB:MakeADFun]{TMB::MakeADFun} as argument \code{map}, over-writing
the version \code{tinyVAST(...)$tmb_inputs$tmb_map} and allowing detailed control
over estimated parameters (advanced feature)}

\item{tmb_random}{input passed to \link[TMB:MakeADFun]{TMB::MakeADFun} as argument \code{random}, over-writing
the version \code{tinyVAST(...)$tmb_inputs$tmb_random} and allowing detailed control
over parameters treated as random effects.  \code{tmb_random = NULL} uses
the default (internal) construction, and use \code{tmb_random = c("empty")}
(or some other name that doesn't match actual parameters) to use
penalized likelihood (presumably while also modifying \code{tmb_map} and \code{tmb_par})}

\item{gmrf_parameterization}{Parameterization to use for the Gaussian Markov
random field, where the default \code{separable} constructs a full-rank and
separable precision matrix, and the alternative \code{projection} constructs
a full-rank and IID precision for variables over time, and then projects
this using the inverse-cholesky of the precision, where this projection
allows for rank-deficient covariance.}

\item{reml}{Logical: use REML (restricted maximum likelihood) estimation rather than
maximum likelihood? Internally, this adds the fixed effects to the
list of random effects to integrate over.}

\item{getJointPrecision}{whether to get the joint precision matrix.  Passed
to \code{\link[TMB]{sdreport}}.}

\item{calculate_deviance_explained}{whether to calculate proportion of deviance
explained.  See \code{\link[=deviance_explained]{deviance_explained()}}}

\item{run_model}{whether to run the model of export TMB objects prior to compilation
(useful for debugging)}

\item{optimizer}{which gradient-based optimizer to use.
"L-BFGS-B" is necessary for penalized likelihood involving >40,000
coefficients.}

\item{suppress_user_warnings}{whether to suppress warnings from
package author regarding dangerous or non-standard options}

\item{get_rsr}{Experimental option, whether to report restricted spatial
regression (RSR) adjusted estimator for covariate responses}

\item{extra_reporting}{Whether to report a much larger set of quantities via
\code{obj$env$report()}}

\item{use_anisotropy}{Whether to estimate two parameters representing
geometric anisotropy}

\item{sar_adjacency}{Whether to use queen or rook adjacency when defining
a Simultaneous Autoregressive spatial precision from a sfc_GEOMETRY
(default is queen)}

\item{nearest_neighbors}{Number of nearest neighbors to use for nearest-neighbors
Gaussian process.  Only used if \code{spatial_domain} is an areal model, and
\code{nearest_neighbors = NULL} reverts to using a SAR model}

\item{barrier_stiffness}{The ratio of local stiffness (the scale of diffusion
rate and resulting decorrelation distance) for barriers relative to normal areas
in the SPDE method when using \code{add_mesh_covariates}.  The
default \code{barrier_stiffness = 0.01} is the value from Bakka et al.
2019.}
}
\value{
An object (list) of class \code{tinyVASTcontrol}, containing either default or
updated values supplied by the user for model settings
}
\description{
Control parameters for tinyVAST
}
\references{
Bakka, H., Vanhatalo, J., Illian, J., Simpson, D., Rue, H. (2019).  Non-stationary Gaussian models with physical barriers. Spatial Statistics, 29, 268-288. \doi{10.1016/j.spasta.2019.01.002}
}
