% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{tinyVASTcontrol}
\alias{tinyVASTcontrol}
\title{Control parameters for tinyVAST}
\usage{
tinyVASTcontrol(
  nlminb_loops = 1,
  newton_loops = 0,
  eval.max = 1000,
  iter.max = 1000,
  getsd = TRUE,
  silent = getOption("tinyVAST.silent", TRUE),
  trace = getOption("tinyVAST.trace", 0),
  verbose = getOption("tinyVAST.verbose", FALSE),
  profile = c(),
  tmb_par = NULL,
  gmrf_parameterization = c("separable", "projection"),
  estimate_delta0 = FALSE,
  getJointPrecision = FALSE,
  calculate_deviance_explained = TRUE
)
}
\arguments{
\item{nlminb_loops}{Integer number of times to call \code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{newton_loops}{Integer number of Newton steps to do after running
\code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{eval.max}{Maximum number of evaluations of the objective function
allowed. Passed to \code{control} in \code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{iter.max}{Maximum number of iterations allowed. Passed to \code{control} in
\code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{getsd}{Boolean indicating whether to call \code{\link[TMB:sdreport]{TMB::sdreport()}}}

\item{silent}{Disable terminal output for inner optimizer?}

\item{trace}{Parameter values are printed every \code{trace} iteration
for the outer optimizer. Passed to
\code{control} in \code{\link[stats:nlminb]{stats::nlminb()}}.}

\item{verbose}{Output additional messages about model steps during fitting?}

\item{profile}{Parameters to profile out of the likelihood (this subset will be appended to \code{random} with Laplace approximation disabled).}

\item{tmb_par}{list of parameters for starting values, with shape identical
to \code{tinyVAST(...)$internal$parlist}}

\item{gmrf_parameterization}{Parameterization to use for the Gaussian Markov
random field, where the default \code{separable} constructs a full-rank and
separable precision matrix, and the alternative \code{projection} constructs
a full-rank and IID precision for variables over time, and then projects
this using the inverse-cholesky of the precision, where this projection
allows for rank-deficient covariance.}

\item{estimate_delta0}{Estimate a delta model?}

\item{getJointPrecision}{whether to get the joint precision matrix.  Passed
to \code{\link[TMB]{sdreport}}.}

\item{calculate_deviance_explained}{whether to calculate proportion of deviance
explained.  See \code{\link[=deviance_explained]{deviance_explained()}}}
}
\description{
Control parameters for tinyVAST
}
