% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{tinyVASTcontrol}
\alias{tinyVASTcontrol}
\title{Control parameters for tinyVAST}
\usage{
tinyVASTcontrol(
  nlminb_loops = 1,
  newton_loops = 0,
  eval.max = 1000,
  iter.max = 1000,
  getsd = TRUE,
  quiet = FALSE,
  trace = 1,
  profile = c(),
  tmb_par = NULL
)
}
\arguments{
\item{newton_loops}{Integer number of newton steps to do after running \code{nlminb}}

\item{getsd}{Boolean indicating whether to call \code{\link[TMB]{sdreport}}}

\item{profile}{Parameters to profile out of the likelihood (this subset will be appended to \code{random} with Laplace approximation disabled).}

\item{tmb_par}{list of parameters for starting values, with shape identical to
\code{fit(...)$internal$parlist}}
}
\description{
Control parameters for tinyVAST
}
