% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{vcov.tinyVAST}
\alias{vcov.tinyVAST}
\title{Extract Variance-Covariance Matrix}
\usage{
\method{vcov}{tinyVAST}(object, which = c("fixed", "random", "both"), ...)
}
\arguments{
\item{object}{output from \code{fit}}

\item{which}{whether to extract the covariance among fixed effects, random effects, or both}

\item{...}{ignored, for method compatibility}
}
\description{
extract the covariance of fixed effects, or both fixed and random effects.
}
